Clean entities in `ClientBuffers` properly
I noticed that I did it partially in the previous PR.

Since it's in the same struct alongside with `info` which I reset before
the use, I decided to use the same approach. Plus I insert data in 4
places and pop it only in one place, so it's less bug-prone to clean in
that one place. Also clearing in `drain_entities` would be odd.