Abstract I/O
With the current implementation in order to disconnect user need to
write library-specific code. For example, in case of renet user will
need to call `RenetClient::disconnect` or remove `RenetClient` resource.

I think we can do better by hiding `RepliconServer::set_active` and
`RepliconClient::set_status` and providing events for messaging
libraries and users to control it. But I will probably do it in a
follow-up PR.

We can include more library-independent things into `RepliconServer`,
but I would also do it in a follow-up PRs.