Make marker API more safe
In this PR I decided to rename `SerdeFns` into `RuleFns` for clarity.
Now it's clear that serialization and deserialization is managed by
rules.

To make the API more safe, I needed `WriteFn` to have an associated
type. To achieve this, I made `RuleFns` typed. In order to store it inside
`ReplicationFns`, I created `UntypedRuleFns` which is similar to the
original one.

Since `WriteFn` also needs to be stored deep inside `ReplicationFns`,
I created a similar `UntypedCommandFns` to store it too.

For organization purposes, I put in `CommandFns` only two functions -
`remove` and `write`. Rest of the fields now stored inside `ComponentFns`.

In order to call a typed version of `WriteFn`, I created a special
`UntypedWriteFn` that will convert the type. It's similar to what
`ReadFn` does. And for consistency I renamed `ReadFn` into `UntypedSerializeFn`.

These changes makes public API for registering command functions safe
and type-checked. And as a small bonus, user now pass `RuleFns` and
`CommandFns` instead of plain function. This way it's much easier to
construct default functions and `deserialize_in_place` can be assigned
via builder pattern.