Provide context struct for all functions
Writing shares context with deserialize since deserialization also need
access to commands for in-place deserialization for components with
entities.
Removal and despawn functions also share context due to their nature.
But it makes sense to keep serialization context different since the
tick is also different and different kind of data can be passed in the
future.

This change have the following benefits:
- We can add more context-related data without breaking changes.
- Reduces the number of arguments in the writing signature.
- More flexible, now serialization logic can depend on tick.
- Removes the need of ugly creation of `ClientMapper` in writing
  functions, now writing context implement `MapEntities`.

Was suggested by @NiseVoid.

I used more specific naming in `client` module to disambiguate between
the current tick and message tick.