Use separate resource on client to track last received init tick
I noticed that people sometimes mess with
`RepliconTick` on client and it can result in broken updates. This is
why I introduced `ServerInitTick` (similar to `ServerEntityTicks`).
It wraps `RepliconTick` and forbids incrementing it.
So users can now also have their own `RepliconTick` on the client for
rollback scenarios where the client has its own tick that is ahead
than the server's tick.