Implement history switch for markers (#247)
We currently discard old updates. But when users implement
interpolation, they need to store all intermediate values.
This PR adds a special field for markers that disables the mentioned
discard.

I considered making it a function property, but I think it's more
ergonomic as a marker property and faster to calculate.

We replaced `AppMarkerExt::register_marker_with_priority` with
`AppMarkerExt::register_marker_with` that also accepts `need_history`
boolean. Not quite sure if it should be a bool, maybe use an enum?

Since some markers need history and some not, we introduced `consume`
function that can be optionally overriden. By default it just disables
mapping entities and calls `deserialize`, but user can optimize it.

Now internally used entity markers is not just a `Vec<bool>`, but a special
struct that also pre-computes if an entity have any marker that requires history.

When we read updates, we calculate it first and then and if current
entity have any marker that needs history, we don't discard it as
before.

Then if an entity is old, we internally invoke newly added
`consume_or_write` that calls `write` if history is required by its
marker or `consume` if not.

---------

Co-authored-by: UkoeHB <37489173+UkoeHB@users.noreply.github.com>