Separate remove context from despawn (#255)
Sometimes it's useful to access `EntityMut` in removal function. Since I
already fetch it on replicon side, I just pass it instead of
`EntityCommands`.

Also I moved commands to context for consistency with other contexts.
This requires to separate remove and despawn contexts, but I think that
it's even better.