Fix replication with visibility policy != `VisibilityPolicy::All`
Visibility was coupled with a check if a client was just connected.
This resulted in a passed test that checks replication of previously
spawned entities (it was running with `VisibilityPolicy::All`). But for
all other policies the logic was incorrect and led to a crash with
"entity should be present after adding component".

I decoupled the logic. We now determine if the entity is new to a
client by checking if it has a tick.

I thinking that we need to rename "change limit" into "init tick" because
we use this terminology on client. But it will be a breaking change, so
I will leave it for the next release.