Add a dedicated readme for renet integration
The separate readme will be displayed on crates.io for
`bevy_replicon_renet` instead of the full replicon readme which is
nicer.

This way we can also provide a link to it from the main readme and
maintain a better compatibility table for it.

Inspired by
https://github.com/Henauxg/bevy_quinnet/tree/main/bevy_replicon_quinnet

I also used shorter notation for GitHub links and docs.rs.