Construct event systems dynamically (#395)
To avoid creating several systems for each event type, we store the IDs of all event resources and process them at once.
This is similar to what Bevy does.
However, since we need access to resources by their IDs, we used exclusive systems with a lot of `resource_scope`.
Now, we can utilize `FilteredResources` and `FilteredResourcesMut` to define the required access and get rid of exclusive systems!