Implement remote triggers (#398)
I decided to implement remote triggers on top of events to reuse existing logic.

Unlike events, we can't simply observe and re-trigger, as we need
access to all targets. To address this, we provide a special `RemoteTrigger` and
extension traits for sending these events internally. Both global and targeted
triggers are supported. But we don't support component-based targets yet, only entity-based.

To allow users to override serialization without manually writing it for
for event targets, I replaced the bare serialization functions with an abstraction
similar to component rules. Additionally, I updated how `ClientEvent` is created
internally to cleanly reuse this logic for triggers.

For API examples see documentation changes in `lib.rs`.