Improve connection triggers
- Split enum into 2 separate triggers. It makes them nicer to observe.
- Replace string-based `reason` with `DisconnectReason` enum.

I expect resources as entities land with Bevy 0.16 and planning to utilize `DisconnectReason` for `RepliconClientStatus::Disconnected` as well.
Right now I just can't provide a convenient API for backends to set it.
But with resources as entities backends will be able to utilize triggers to react on removals and update the `RepliconClientStatus` nicely.