Store connected clients and their data as entities (#423)
This PR replaces `ConnectedEntities`, `ReplicatedEntities`, `ClientEntityMap`, and `Messages` resources with entities and data stored as components.

Connected clients are now represented by entities with a `ConnectedEntity` component. Connected entities become replicated after the insertion of `ReplicatedClient` (which replaces the `StartReplication` trigger). It's just a marker now, I've split the original `ReplicatedClient` into multiple components. Visibility and statistics are now accessible via the `ClientVisibility` and `ClientStats` components on replicated entities. The `ClientEntityMap` is now a component on replicated entities that stores mappings for them. `ClientTicks` is no longer accessible to users, but it isn't needed for the public API - even prediction crates don't use it.

`UpdateMessage` and `MutateMessages` are now crate-private components, just to speed up iteration. Using `zip` is slower, but if everything is on an entity, the performance is similar to what we had before this PR. It's still a bit slower, but I believe the improved ergonomics are worth it.

Since connected clients now have a unique identifier (`Entity`). This also gives us both fast iteration and fast lookup (closes #326). Users will mostly interact with entities, and plugins can use `ClientId` if they need a persistent identifier.

The server ID is now `SERVER` constant which equal to `Entity::PLACEHOLDER`.

I also removed the `ClientConnected` and `ClientDisconnected` events. Users can observe for `Trigger<OnAdd, ConnectedClient>` or `Trigger<OnRemove, ConnectedClient>`. `DisconnectReason` is also gone. We don't care about it in Replicon and can't provide a nice representation for it anyway. Users can access proper information from the used backend. There is no need to duplicate it for us.