Move `ConnectedClient::id` into a separate component
Some backends don't support persistent identifiers. To avoid forcing them to use bytes from `Entity` for this, `ConnectedClient::id` is now a separate component.

This makes it less explicit, but I included information about this in the documentation, so it shouldn't be easy to miss.
I think this way `ConnectedClient` looks cleaner (no `new` method needed) and should integrate better with possible component indexing in the future.

I also renamed it from `ClientId` to `NetworkId` to avoid confusion with client entities (which are also technically client IDs).