Make channels less opinionated (#447)
* Make channels less opinionated

We require only specific delivery guarantees. All other configuration
can be done on the backend side. It doesn't make sense to provide
settings we don't care about and that some backends might ignore.

To provide a way to configure channels for remote events, I exposed
`EventRegistry`, which provides channel IDs that users can use to
configure channels on the backend side. I had to rename the registry to
`RemoteEventRegistry` since Bevy already has a resource with this name.

We also no longer limit the number of channels and delegate this to the
backend.

And since we no longer have `RepliconChannel`, I renamed `ChannelKind`
to just `Channel`.

* Rename `channels` module into `replicon_channels`

Because it contains `RepliconChannels` with just a few enums.