Process all received data with a disconnect
This should be handled by the messaging backend.
Instead of marking the client as disconnected immediately,
backends should defer this until `PostUpdate`.

I attempted to implement this on Replicon's side in #493,
but this approach forces users to use
`client_just_disconnected.or(client_connected)`
conditions to receive messages on disconnect,
which is error-prone.

I've documented this behavior and added tests for the example backend.